+++ src/Dictionary.test.ts
@@ -1,11 +1,11 @@
 import { HashMap } from "@rimbu/hashed";
-import { BasicDictionary } from "./BasicDictionary";
+import { Dictionary } from "./Dictionary";
 
-describe("Basic dictionary", () => {
+describe("Dictionary", () => {
   describe("when requesting a single-word translation", () => {
     describe("when the translation is available", () => {
       it("should return the translation", () => {
-        const dictionary = new BasicDictionary(HashMap.of(["dodo", "ciop"]));
+        const dictionary = new Dictionary(HashMap.of(["dodo", "ciop"]));
         const translation = dictionary.translate("dodo");
 
         expect(translation).toBe("ciop");
@@ -14,7 +14,7 @@ describe("Basic dictionary", () => {
 
     describe("when the translation is NOT available", () => {
       it("should return the phrase itself", () => {
-        const dictionary = new BasicDictionary();
+        const dictionary = new Dictionary();
         const translation = dictionary.translate("dodo");
 
         expect(translation).toBe("dodo");
@@ -25,7 +25,7 @@ describe("Basic dictionary", () => {
   describe("when requesting a multi-word translation", () => {
     describe("when the translation is available", () => {
       it("should return the translation", () => {
-        const dictionary = new BasicDictionary(
+        const dictionary = new Dictionary(
           HashMap.of(["a b c", "alpha beta gamma"])
         );
         const translation = dictionary.translate("a b c");
@@ -36,7 +36,7 @@ describe("Basic dictionary", () => {
 
     describe("when the translation is NOT available", () => {
       it("should return the phrase itself", () => {
-        const dictionary = new BasicDictionary();
+        const dictionary = new Dictionary();
         const translation = dictionary.translate("a b c");
 
         expect(translation).toBe("a b c");
@@ -47,7 +47,7 @@ describe("Basic dictionary", () => {
   describe("when requesting raw translations", () => {
     describe("when the dictionary is empty", () => {
       it("should return an empty object", () => {
-        const dictionary = new BasicDictionary();
+        const dictionary = new Dictionary();
         const rawTranslations = dictionary.toRawTranslations();
 
         expect(rawTranslations).toBeEmptyObject();
@@ -56,7 +56,7 @@ describe("Basic dictionary", () => {
 
     describe("when the dictionary contains translations", () => {
       it("should return all the translations", () => {
-        const dictionary = new BasicDictionary(
+        const dictionary = new Dictionary(
           HashMap.of(["cip", "ciop"], ["yogi", "bubu"])
         );
         const rawTranslations = dictionary.toRawTranslations();